mb/google/peppy: add _DSD to touchscreen ACPI
Recent changes to the Atmel touchscreen driver in the mainline
kernel broke functionality with devices running upstream coreboot,
due to assumption that the i2c devices would be in PCI mode
(as with the stock Google firmware) rather than in ACPI mode as
they are in upstream coreboot. Mitigate this by adding the
required property so the driver will correctly attach.

Test: build/boot peppy on 4.18+ kernel, verify touchscreen working

Change-Id: I05df8367886eef55b409590f75a68d98d4e5fbdf
Signed-off-by: Matt DeVillier <matt.devillier@gmail.com>